- add package specification and its dumper - and some helper for loading witch not tested
